==> analysis/2_batch_effect_correction.Rmd <==
---
title: "IMC data analysis workflow"
author: "Tess Brodie"
date: "University of Bern, DBMR, Visceral and Transplantation Surgery lab"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

# Batch effect correction

### Of note: the correction approaches presented here aim at removing any differences between samples. This will also remove biological differences between the patients/indications. Nevertheless, integrating cells across samples can facilitate the detection of cell phenotypes via clustering.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#Important - this path is relative to your dataset, set working directory
knitr::opts_knit$set(root.dir = "C:/Users/IMC Platform laptop/Users2/Tess/SteinbockCellSegmentation/220707/220707_workflowr")

#Set cache
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)

#To make figures
knitr::opts_chunk$set(dev = "png")
```

```{css, echo=FALSE}
.scroll-100 {
  max-height: 100px;
  overflow-y: auto;
  background-color: inherit;
}
```

### Load packages

```{r load-libs, message=FALSE}
library(batchelor)
library(scater)
library(cowplot)
library(dittoSeq)
library(viridis)
library(harmony)
library(Seurat)
library(SeuratObject)

```

### Load data

```{r load spe1}
spe <- readRDS(file.path("output", "spe1.rds"))
```

### FastMNN method to correct for batch effect:

```{r FastMNN sample correction}
out <- fastMNN(spe, batch = spe$patient_id,
               auto.merge = TRUE,
               assay.type = "exprs")

# Transfer the correction results to the main spe object
reducedDim(spe, "fastMNN") <- reducedDim(out, "corrected")
```

### QC of corrected results: gives batch effect and lost variance values if combining samples 

```{r table batch effect and lost.var, class.output="scroll-100"}
merge_info <- metadata(out)$merge.info 

DataFrame(left = merge_info$left,
          right = merge_info$right,
          batch.size = merge_info$batch.size,
          max_lost_var = rowMax(merge_info$lost.var))
```

### Visualization of batch corrections:
### recompute the UMAP embedding using the corrected low-dimensional coordinates for each cell

```{r UMAP with corrected values, class.output="scroll-100"}
spe <- runUMAP(spe, dimred= "fastMNN", name = "UMAP_mnnCorrected") 

# visualize patient id 
p1 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP before correction")
p2 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP_mnnCorrected", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP after correction")

#pdf("UMAP_correctedvalues.pdf", width = 20, height = 9)
plot_grid(p1, p2)
dev.off()
plot_grid(p1, p2)
```

### Visualize the expression of selected markers across all cells before correction

```{r marker expression before correction, class.output="scroll-100"}
markers <- c("E-cadherin", "CD8a", "CD4", "CD3e", "HNFa", "Ki-67", "CD19", "a-SMA", "CD11b", "Ly6-G", "Pan-actin", "CD31", "CD74", "CD44", "Arginase-1", "F4_80", "CD206" )

# Before correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 


plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())


plot_grid(plotlist = plot_list) 
```

### Visualize the expression of selected markers across all cells after correction

```{r marker expression after correction, class.output="scroll-100"}
# After correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP_mnnCorrected", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 
plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())


plot_grid(plotlist = plot_list) 
```

### Harmony correction:
### Harmony returns the corrected low-dimensional coordinates for each cell

```{r harmony correction storage}

mat <- t(assay(spe, "exprs"))

harmony_emb <- HarmonyMatrix(mat, spe$patient_id, do_pca = TRUE, npcs = 15)

reducedDim(spe, "UMAP_harmony") <- harmony_emb
```

### Visualize the cells in low dimensions after UMAP embedding:

```{r UMAP before vs. after harmony correction}
# visualize patient id 
p1 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP before correction")
p2 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP_harmony", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP after correction")

#pdf("UMAP_harmonycorrection.pdf", width = 20, height = 9)
plot_grid(p1, p2)
dev.off()

plot_grid(p1, p2)
```

### Visualize the expression of selected markers across all cells before correction

```{r UMAP_before, class.output="scroll-100"}
# Before correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 
plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())


plot_grid(plotlist = plot_list) 
```

### Visualize the expression of selected markers across all cells after correction

```{r UMAP_after, class.output="scroll-100"}
# After correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP_harmony", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 
plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())


plot_grid(plotlist = plot_list) 
```

### Seurat correction
### Seurat performs batch correction similarly to fastMNN by finding mutual nearest neighbors (MNN) in low dimensional space before correcting the expression values of cells

```{r create Seurat object, eval=FALSE, include=FALSE}
seurat_obj <- as.Seurat(spe, counts = "counts", data = "exprs")
seurat_obj <- AddMetaData(seurat_obj, as.data.frame(colData(spe)))

seurat.list <- SplitObject(seurat_obj, split.by = "patient_id")
```

## Integrate markers and patient Ids ### Did not work for me#

```{r integratoins by seurat, eval=FALSE, include=FALSE}
features <- rownames(spe)
seurat.list <- lapply(X = seurat.list, FUN = function(x) {
    x <- ScaleData(x, features = features, verbose = FALSE)
    x <- RunPCA(x, features = features, verbose = FALSE, npcs = 19, approx = FALSE)
    return(x)
})

anchors <- FindIntegrationAnchors(object.list = seurat.list, 
                                  anchor.features = features,
                                  reduction = "rpca", 
                                  k.anchor = 19)

combined <- IntegrateData(anchorset = anchors)
```

### PCA reduction in seurat method:

```{r seurat pca, eval=FALSE, include=FALSE}
DefaultAssay(combined) <- "integrated"

combined <- ScaleData(combined, verbose = FALSE)
combined <- RunPCA(combined, npcs = 19, verbose = FALSE)

reducedDim(spe, "seurat") <- combined@reductions$pca@cell.embeddings
```

### Recompute the UMAP embeddings based on Seurat integrated results and visualize the embedding

```{r seurat UMAP, eval=FALSE, include=FALSE}
spe <- runUMAP(spe, dimred = "seurat", name = "UMAP_seurat") 
```

### Visualize patient id 

```{r patient id visualization, eval=FALSE, include=FALSE}
p1 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP before correction")
p2 <- dittoDimPlot(spe, var = "patient_id", 
                   reduction.use = "UMAP_seurat", size = 0.2) + 
    scale_color_manual(values = metadata(spe)$color_vectors$patient_id) +
    ggtitle("Patient ID on UMAP after correction")

plot_grid(p1, p2)
```

### Visualize the expression of selected markers across all cells before correction

```{r, eval=FALSE, include=FALSE}
# Before correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 
plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())
plot_grid(plotlist = plot_list) 
```

### Visualize the expression of selected markers across all cells after correction

```{r, eval=FALSE, include=FALSE}
# After correction
plot_list <- multi_dittoDimPlot(spe, var = markers, reduction.use = "UMAP_seurat", 
                   assay = "exprs", size = 0.2, list.out = TRUE) 
plot_list <- lapply(plot_list, function(x) x + scale_color_viridis())
plot_grid(plotlist = plot_list) 
```

### Save data

```{r save spe2}
saveRDS(spe, file.path("output", "spe2.rds"))
```
