==> analysis/4_cell_annotation.Rmd <==
---
title: "IMC data analysis workflow"
author: "Tess Brodie"
date: "University of Bern, DBMR, Visceral and Transplantation Surgery lab"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---


# Cell classification

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

#Important - this path is relative to your dataset, set working directory
knitr::opts_knit$set(root.dir = "C:/Users/IMC Platform laptop/Users2/Tess/SteinbockCellSegmentation/220707/220707_workflowr")

#Set cache
knitr::opts_chunk$set(autodep = TRUE, cache = TRUE, cache.lazy = FALSE)

#To make figures
knitr::opts_chunk$set(dev = "png")
```

```{css, echo=FALSE}
.scroll-100 {
  max-height: 100px;
  overflow-y: auto;
  background-color: inherit;
}
```

### Load packages 

```{r load-libs, message=FALSE}
library(SpatialExperiment)
library(Rphenograph)
library(igraph)
library(dittoSeq)
library(viridis)
library(bluster)
library(BiocParallel)
library(ggplot2)
library(scran)
library(CATALYST)
library(kohonen)
library(ConsensusClusterPlus)
library(patchwork)
library(pheatmap)
library(gridExtra)
library(cytomapper)
library(Rphenoannoy)
library(SingleCellExperiment)
library(caret)
library(ggridges)
library(patchwork)
library(pheatmap)
library(gridExtra)
library(scuttle)
library(stringr)
library(tidyverse)

```

### Load data

```{r load data}
spe <- readRDS(file.path("output", "spe3.rds"))

# Sample cells
cur_cells <- sample(seq_len(ncol(spe)), 2000)
```

## Cell classification: manual labeling of cells with cytomapper shiny app (not working well due to doublets)

```{r cytomapperShiny, eval=FALSE, include=FALSE}
if (interactive()) {
    
    images <- readRDS("data/images.rds")
    masks <- readRDS("data/masks.rds")
    
    cytomapperShiny(object = spe, mask = masks, image = images, 
                    cell_id = "ObjectNumber", img_id = "sample_id")
}

spe$all <- "all"
cytomapperShiny(object = spe, img_id = "all", cell_id = "ObjectNumber")

```

### Define color vectors

```{r colors, eval=FALSE, include=FALSE}
#celltype <- setNames(c("#3F1B03", "#F4AD31", "#894F36", "#1C750C", "#EF8ECC", 
#                       "#6471E2", "#4DB23B", "grey", "#F4800C"),
#                     c("B cells", "T cells", "Hepatocytes", "Kupffer cells", "LSECs", 
#                       "neutrophils", "CD31 endos", "Cholangiocytes", "CD11b pos no F480 or Ly6G"))

#metadata(spe)$color_vectors$celltype <- celltype
```

### Read in and consolidate labeled data

```{r read in data}
label_files <- list.files("gated_cells2", 
                          full.names = TRUE, pattern = ".rds$")

# Read in SPE objects
spes <- lapply(label_files, readRDS)

# Merge SPE objects
concat_spe <- do.call("cbind", spes)

#clean indices from multiple ROI

#cleaned <- substring(concat_spe$cytomapper_CellLabel, 3)

cleaned <- gsub('.{2}$', '', concat_spe$cytomapper_CellLabel)


#Change K to k in cleaned

#cleaned <- gsub(pattern = "Kupffer", replacement = "kupffer", cleaned)

#Identify doublets

cur_tab <- unclass(table(colnames(concat_spe), cleaned))
                         
colSums(cur_tab)

cur_labels <- rep("doublets", nrow(cur_tab))
names(cur_labels) <- rownames(cur_tab)

# Single assignments
single_index <- rowSums(cur_tab) == 1
cur_labels[single_index] <- colnames(cur_tab)[apply(cur_tab[single_index,], 1, 
                                                    which.max)]

# Check for number of double ID cells
cur_tab[rowSums(cur_tab > 1),]

# Remove doublets
cur_labels <- cur_labels[cur_labels != "doublets"]
table(cur_labels)

# Transfer labels to SPE object
spe_labels <- rep("unlabeled", ncol(spe))
names(spe_labels) <- colnames(spe)
spe_labels[names(cur_labels)] <- cur_labels
spe$cell_labels <- spe_labels

# Number of cells labeled per patient
table(spe$cell_labels, spe$patient_id)

```

### Train classifier

```{r Train classifier}
# Split between labeled and unlabeled cells
lab_spe <- spe[,spe$cell_labels != "unlabeled"]
unlab_spe <- spe[,spe$cell_labels == "unlabeled"]

# Randomly split into train and test data
set.seed(20220718)
trainIndex <- createDataPartition(factor(lab_spe$cell_labels), p = 0.75)
train_spe <- lab_spe[,trainIndex$Resample1]
test_spe <- lab_spe[,-trainIndex$Resample1]

# Specify train parameters for 5-fold cross validation
fitControl <- trainControl(method = "cv",
                           number = 5)

# Select the data for training
cur_mat <- t(assay(train_spe, "exprs"))

# Train a random forest model for predicting cell labels
# This call also performs parameter tuning
#set.seed(220510)
rffit <- train(x = cur_mat, 
               y = factor(train_spe$cell_labels),
               method = "rf", ntree = 1000,
               tuneLength = 5,
               trControl = fitControl)

rffit
```

### Classifier performance

```{r check performance}
ggplot(rffit) + 
  geom_errorbar(data = rffit$results,
                aes(ymin = Accuracy - AccuracySD,
                    ymax = Accuracy + AccuracySD),
                width = 0.4) +
    theme_classic(base_size = 15)

plot(varImp(rffit))

# Select test data
cur_mat <- t(assay(test_spe, "exprs"))

# Predict cell phenotypes in test data
cur_pred <- predict(rffit, 
                    newdata = cur_mat)

cm <- confusionMatrix(data = cur_pred, 
                      reference = factor(test_spe$cell_labels), 
                      mode = "everything")

cm
```

### Visualize classifier performance

```{r visualize performance}
data.frame(cm$byClass) %>%
  mutate(class = sub("Class: ", "", rownames(cm$byClass))) %>%
  ggplot() + 
  geom_point(aes(1 - Specificity, Sensitivity, 
                 size = Detection.Rate,
                 fill = class),
             shape = 21) + 
  #scale_fill_manual(values = metadata(spe)$color_vectors$celltype) +
  theme_classic(base_size = 15) + 
  ylab("Sensitivity (TPR)") +
  xlab("1 - Specificity (FPR)")
```

```{r boxplots, fig.width= 8, fig.height=16}
cur_pred <- predict(rffit, 
                    newdata = cur_mat, 
                    type = "prob")
cur_pred$truth <- factor(test_spe$cell_labels)

cur_pred %>%
  pivot_longer(cols = neutro:B) %>%
  ggplot() +
  geom_boxplot(aes(x = name, y = value, fill = name), outlier.size = 0.5) +
  facet_wrap(. ~ truth, ncol = 1) + 
  #scale_fill_manual(values = metadata(spe)$color_vectors$celltype)  +
  theme(panel.background = element_blank(), 
        axis.text.x = element_text(angle = 45, hjust = 1))
```

### Classification of new data

```{r classify new data}
# Select unlabeled data
cur_mat <- t(assay(unlab_spe, "exprs"))

# Predict cell phenotypes
cell_class <- as.character(predict.train(rffit, 
                                         newdata = cur_mat, 
                                         type = "raw"))
names(cell_class) <- rownames(cur_mat)

# Extract classification probabilities
cell_prob <- predict.train(rffit, 
                           newdata = cur_mat, 
                           type = "prob")


# Distribution of maximum probabilities
tibble(max_prob = rowMax(as.matrix(cell_prob)),
       type = cell_class) %>%
    ggplot() +
        geom_density_ridges(aes(x = max_prob, y = cell_class, fill = cell_class)) +
        #scale_fill_manual(values = metadata(spe)$color_vectors$celltype) +
        theme_classic(base_size = 15) +
        xlab("Maximum probability") +
        ylab("Cell type") + 
        xlim(c(0,1.2))

# Label undefined cells
cell_class[rowMax(as.matrix(cell_prob)) < 0.4] <- "undefined"

# Store labels in SpatialExperiment onject
cell_labels <- spe$cell_labels
cell_labels[colnames(unlab_spe)] <- cell_class
spe$celltype <- cell_labels 

table(spe$celltype, spe$patient_id)
```

### Visualize comparison of cell label classification with clusters using heatmaps

```{r heatmap}
tab1 <- table(spe$celltype, 
              paste("Rphenograph", spe$pg_clusters))
tab2 <- table(spe$celltype, 
              paste("SNN", spe$nn_clusters))
tab3 <- table(spe$celltype, 
              paste("SOM", spe$som_clusters))

pheatmap(log10(tab1 + 10), color = viridis(100))

pheatmap(log10(tab2 + 10), color = viridis(100))

pheatmap(log10(tab3 + 10), color = viridis(100))

tab1 <- table(spe$celltype, 
              paste("Rphenograph", spe$pg_clusters_corrected))
tab2 <- table(spe$celltype, 
              paste("SNN", spe$nn_clusters_corrected))
tab3 <- table(spe$celltype, 
              paste("SOM", spe$som_clusters_corrected))

pheatmap(log10(tab2 + 10), color = viridis(100))

pheatmap(log10(tab3 + 10), color = viridis(100))

pheatmap(log10(tab1 + 10), color = viridis(100))
```

### Save data

```{r save spe4}
saveRDS(spe, file.path("output", "spe4.rds"))
```
